refactor: Update base.html to include custom CSS file
The base.html template has been updated to include a custom CSS file called "style.css". This file is located in the "static/css" directory and is now linked in the head section of the HTML. This change allows for custom styling to be applied to the web pages.

Note: This commit message follows the established commit message conventions in the repository.