feat: Update home.html to use anchor tag for clear button
The home.html template has been modified to use an anchor tag instead of an input button for the "CLEAR" functionality. This change improves the accessibility and usability of the button by providing a clickable link. The anchor tag is styled using the "btn" and "btn-warning" classes.

Note: This commit message follows the established commit message conventions in the repository.